Remove events module
